refactor: Add explicit dtype conversion for IoU-based confidence targets to prevent type mismatch
- Cast pred_ious to conf_target.dtype before assignment to ensure type compatibility
- Prevents potential float16/float32 mismatch when mixed precision training is enabled
- Maintains IoU-aware confidence targets from previous refactor (dbbf5a4)

+++ models/yowo/loss_multitask.py
@@ -414,7 +414,7 @@ class MultiTaskCriterion(object):
                 num_anchors = sum([ab.shape[0] for ab in anchors])
                 conf_target = conf_preds.new_zeros((num_anchors, 1))
                 # Foreground anchors get IoU as target (quality-aware)
-                conf_target[fg_mask] = pred_ious.unsqueeze(-1)
+                conf_target[fg_mask] = pred_ious.unsqueeze(-1).to(conf_target.dtype)
                 
                 box_target = tgt_bboxes_scaled[matched_gt_inds]
